server/pyproject.toml
[project]
name = "app-scaffold"
version = "0.1.0"
description = "Full-stack scaffold — FastAPI + React + PostgreSQL + Google ADK"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "pydantic[email]>=2.7",
  "pydantic-settings>=2.3",
  "sqlalchemy>=2.0",
  "asyncpg>=0.30",
  "bcrypt>=4.0",
  "pyjwt[crypto]>=2.8",
  "python-multipart>=0.0.9",
  "alembic>=1.14",
  "loguru>=0.7",
  "google-adk>=1.0",
  "pgvector>=0.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "ruff>=0.5",
  "mypy>=1.10",
  "import-linter>=2.0",
]

[tool.uv]
managed = true

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests/unit", "tests/integration"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
exclude = ["alembic/"]

[tool.importlinter]
root_packages = ["app"]

[[tool.importlinter.contracts]]
name = "Services do not import from API"
type = "forbidden"
source_modules = ["app.services"]
forbidden_modules = ["app.api"]

[[tool.importlinter.contracts]]
name = "DB does not import from services or API"
type = "forbidden"
source_modules = ["app.db"]
forbidden_modules = ["app.services", "app.api"]

[[tool.importlinter.contracts]]
name = "Agent does not import from API or services"
type = "forbidden"
source_modules = ["app.agent"]
forbidden_modules = ["app.api", "app.services"]
